table creation, navbar imported
