+++ frontend/src/App.js
@@ -6,7 +6,7 @@ import LoginPage from "./pages/LoginPage";
 import RegisterPage from "./pages/RegisterPage";
 //import Events from "./components/Events";
 import StudentPage from "./components/StudentPage";
-import AdminPage from "./components/AdminPage";
+import AdminPage from "./pages/AdminPage";
 import CompanyPage from "./components/CompanyPage";
 
 function App() {
