+++ frontend/src/Module/EventManagement.module.css
@@ -1,0 +1,3 @@
+.button {
+    width: 150px; 
+  }
