+++ frontend/src/pages/AdminPage.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+import Navbar from "../components/Navbar";
+import EventManagement from "../components/EventManagement";
+
+const AdminPage = () => {
+    return (
+        <div>
+            <Navbar />
+            <EventManagement />
+        </div>
+    )
+}
+
+export default AdminPage;
